// frontend/src/api-ms-cv-v2/js-client/api.ts
export interface CompetenceResponseDTO {
    'nom'?: string;
    'niveau'?: string;
}
// ...
export interface CvResponseDTO {
    'id'?: number;
    'userId'?: number;
    'prenom'?: string;
    'nom'?: string;
    'templateId'?: number;
    'templateNom'?: string;
    'titreProfil'?: string;
    'telephone'?: string;
    'email'?: string;
    'adresse'?: string;
    'resumeProfil'?: string;
    'createdAt'?: string;
    'updatedAt'?: string;
    'formations'?: Array<FormationResponseDTO>;
    'experiences'?: Array<ExperienceResponseDTO>;
    'competences'?: Array<CompetenceResponseDTO>;
    'langues'?: Array<LangueResponseDTO>;
    'interets'?: Array<InteretResponseDTO>;
}
export interface ExperienceDTO {
    'poste'?: string;
    'entreprise'?: string;
    'ville'?: string;
    'dateDebut'?: string;
    'dateFin'?: string;
    'description'?: string;
}
export interface ExperienceResponseDTO {
    'poste'?: string;
    'entreprise'?: string;
    'ville'?: string;
    'dateDebut'?: string;
    'dateFin'?: string;
    'description'?: string;
}
export interface FormationDTO {
    'diplome'?: string;
    'etablissement'?: string;
    'ville'?: string;
    'anneeDebut'?: string;
    'anneeFin'?: string;
}
export interface FormationResponseDTO {
    'diplome'?: string;
    'etablissement'?: string;
    'ville'?: string;
    'anneeDebut'?: string;
    'anneeFin'?: string;
}
export interface InteretDTO {
    'libelle'?: string;
}
export interface InteretResponseDTO {
    'libelle'?: string;
}
export interface LangueDTO {
    'nom'?: string;
    'niveau'?: string;
}
export interface LangueResponseDTO {
    'nom'?: string;
    'niveau'?: string;
}